services/app/requirements.txt
uvicorn
fastapi
celery
redis
ayv
pydantic-settings
python-dotenv